codex: carve wire.rs (pure move)
JSON-RPC wire types, pending-approval bookkeeping, and the inherited
MCP-server config suppression; 14 tests ride along. Struct fields
widened pub(crate) -- mod.rs and the reader construct and read them
across the new module boundary.
